Divided class on different methods + renamed productsService
